refactor: decouple runtime and strategy orchestration
