+++ backend/src/agentv2/runtime/agentRuntime.ts
@@ -8,50 +8,29 @@ import {
   createIterationStrategyPlanner,
   type IterationStrategyPlanner,
 } from '../../agent/agents/iterationStrategyPlanner';
-import type { Intent, ReferencedEntity, StreamingUpdate } from '../../agent/types';
+import type { Intent, StreamingUpdate } from '../../agent/types';
 import type { Finding } from '../../agent/types';
 import { FocusStore, type FocusInteraction } from '../../agent/context/focusStore';
 import {
   AnalysisOptions,
   AnalysisResult,
   AgentRuntimeConfig,
-  AnalysisServices,
   DEFAULT_CONFIG,
   ExecutionContext,
-  ExecutorResult,
   ProgressEmitter,
 } from '../../agent/core/orchestratorTypes';
-import { sessionContextManager, type EnhancedSessionContext } from '../../agent/context/enhancedSessionContext';
-import { understandIntent } from '../../agent/core/intentUnderstanding';
-import { generateInitialHypotheses } from '../../agent/core/hypothesisGenerator';
-import { resolveFollowUp, type FollowUpResolution } from '../../agent/core/followUpHandler';
-import { resolveDrillDown } from '../../agent/core/drillDownResolver';
-import type { FocusInterval } from '../../agent/strategies/types';
+import { sessionContextManager } from '../../agent/context/enhancedSessionContext';
 import {
   createEnhancedStrategyRegistry,
-  type StrategyMatchResult,
   type StrategyRegistry,
 } from '../../agent/strategies';
-import { createAgentMessageBus } from '../../agent/communication';
 import { CircuitBreaker } from '../../agent/core/circuitBreaker';
-import { createEmittedEnvelopeRegistry } from '../../agent/core/emittedEnvelopeRegistry';
-import { detectTraceContext } from '../../agent/core/strategySelector';
-import type { AnalysisExecutor } from '../../agent/core/executors/analysisExecutor';
-import { DirectDrillDownExecutor } from '../operations/directDrillDownExecutor';
-import { StrategyExecutor } from '../../agent/core/executors/strategyExecutor';
-import { HypothesisExecutor } from '../../agent/core/executors/hypothesisExecutor';
-import { applyCapturedEntities } from '../../agent/core/entityCapture';
-import { deriveConclusionContract, generateConclusion } from '../../agent/core/conclusionGenerator';
-import { resolveConclusionScene } from '../../agent/core/conclusionSceneTemplates';
-import { DEEP_REASON_LABEL } from '../../utils/analysisNarrative';
 import {
   IncrementalAnalyzer,
   type IncrementalScope,
   type PreviousAnalysisState,
 } from '../../agent/core/incrementalAnalyzer';
 import { InterventionController } from '../../agent/core/interventionController';
-import { ComparisonExecutor } from '../operations/comparisonExecutor';
-import { ExtendExecutor } from '../operations/extendExecutor';
 import {
   DecisionContext,
   PrincipleDecision,
@@ -64,7 +43,15 @@ import { ApprovalController } from '../operations/approvalController';
 import { PrincipleEngine } from '../principles/principleEngine';
 import { createSoulProfile } from '../soul/soulProfile';
 import { evaluateSoulGuard } from '../soul/soulGuard';
-import { shouldPreferHypothesisLoop } from '../../agent/config/domainManifest';
+import {
+  buildNativeClarifyFallback,
+  buildNativeClarifyPrompt,
+  prepareRuntimeContext,
+  type PreparedRuntimeContext,
+} from './runtimeContextBuilder';
+import { RuntimeExecutionFactory } from './runtimeExecutionFactory';
+import { RuntimeResultFinalizer } from './runtimeResultFinalizer';
+import { selectInitialExecutor } from './runtimeInitialExecutorSelector';
 
 export type AgentRuntimeAnalysisResult = AnalysisResult;
 
@@ -82,18 +69,8 @@ export class AgentRuntime extends EventEmitter {
   private readonly circuitBreaker: CircuitBreaker;
   private readonly focusStore: FocusStore;
   private readonly interventionController: InterventionController;
-
-  private resolveSceneIdHint(intent: Intent, findings: Finding[]): string | undefined {
-    try {
-      return resolveConclusionScene({
-        intent,
-        findings,
-        deepReasonLabel: DEEP_REASON_LABEL,
-      }).selectedTemplate.id;
-    } catch {
-      return undefined;
-    }
-  }
+  private readonly executionFactory: RuntimeExecutionFactory;
+  private readonly resultFinalizer: RuntimeResultFinalizer;
 
   constructor(modelRouter: ModelRouter, config?: Partial<AgentRuntimeConfig>) {
     super();
@@ -120,6 +97,16 @@ export class AgentRuntime extends EventEmitter {
     this.operationExecutor = new OperationExecutor(
       new ApprovalController(this.interventionController)
     );
+
+    this.executionFactory = new RuntimeExecutionFactory({
+      modelRouter: this.modelRouter,
+      runtimeConfig: this.runtimeConfig,
+      agentRegistry: this.agentRegistry,
+      circuitBreaker: this.circuitBreaker,
+      focusStore: this.focusStore,
+    });
+    this.resultFinalizer = new RuntimeResultFinalizer(this.modelRouter, this.interventionController);
+
     this.setupInterventionEventForwarding();
   }
 
@@ -130,7 +117,14 @@ export class AgentRuntime extends EventEmitter {
     options: AnalysisOptions = {}
   ): Promise<AgentRuntimeAnalysisResult> {
     const sessionContext = sessionContextManager.getOrCreate(sessionId, traceId);
-    const runtimeContext = await this.prepareRuntimeContext(query, sessionContext, options);
+    const runtimeContext = await prepareRuntimeContext({
+      query,
+      sessionContext,
+      options,
+      modelRouter: this.modelRouter,
+      emitter: this.createRuntimeEmitter(),
+    });
+
     const decision = this.principleEngine.decide(runtimeContext.decisionContext);
     const plan = this.planner.buildPlan({ context: runtimeContext.decisionContext, policy: decision.policy });
 
@@ -218,48 +212,6 @@ export class AgentRuntime extends EventEmitter {
     this.focusStore.clear();
   }
 
-  private async prepareRuntimeContext(
-    query: string,
-    sessionContext: EnhancedSessionContext,
-    options: AnalysisOptions
-  ): Promise<PreparedRuntimeContext> {
-    const emitter = this.createRuntimeEmitter();
-    const intent = await understandIntent(query, sessionContext, this.modelRouter, emitter);
-    const followUp = resolveFollowUp(intent, sessionContext);
-
-    let drillDownIntervals = followUp.focusIntervals;
-    if (intent.followUpType === 'drill_down') {
-      const drillResolved = await resolveDrillDown(
-        intent,
-        followUp,
-        sessionContext,
-        options.traceProcessorService,
-        sessionContext.getTraceId()
-      );
-      if (drillResolved?.intervals.length) {
-        drillDownIntervals = drillResolved.intervals;
-      }
-    }
-
-    const decisionContext = buildDecisionContextFromIntent(
-      query,
-      sessionContext,
-      intent,
-      followUp,
-      drillDownIntervals || []
-    );
-
-    const executionOptions = buildRuntimeExecutionOptions(options, followUp, drillDownIntervals, intent);
-
-    return {
-      sessionContext,
-      intent,
-      followUp,
-      decisionContext,
-      executionOptions,
-    };
-  }
-
   private async executeWithRuntimeMode(
     runtimeContext: PreparedRuntimeContext,
     query: string,
@@ -285,62 +237,9 @@ export class AgentRuntime extends EventEmitter {
     traceId: string,
     runtimeContext: PreparedRuntimeContext
   ): Promise<AgentRuntimeAnalysisResult> {
-    const services = this.createNativeExecutionServices();
+    const services = this.executionFactory.createExecutionServices();
     const emitter = this.createRuntimeEmitter();
     const sharedContext = services.messageBus.createSharedContext(sessionId, traceId);
-
-    this.strategyPlanner.resetProgressTracking();
-
-    const traceAgentState = runtimeContext.sessionContext.getOrCreateTraceAgentState(query);
-    const hardMaxRounds = Math.max(1, this.runtimeConfig.maxRounds);
-    const softMaxRounds = Math.max(
-      1,
-      Math.floor(Number(traceAgentState.preferences?.maxExperimentsPerTurn || 3))
-    );
-    const effectiveConfig: AgentRuntimeConfig = {
-      ...this.runtimeConfig,
-      maxRounds: hardMaxRounds,
-      softMaxRounds: Math.min(hardMaxRounds, softMaxRounds),
-    };
-
-    const initialHypotheses = await generateInitialHypotheses(
-      query,
-      runtimeContext.intent,
-      runtimeContext.sessionContext,
-      this.modelRouter,
-      this.agentRegistry,
-      emitter
-    );
-    for (const hypothesis of initialHypotheses) {
-      services.messageBus.updateHypothesis(hypothesis);
-    }
-
-    let strategyMatchResult: StrategyMatchResult | null = null;
-    try {
-      const traceContext = runtimeContext.executionOptions.traceProcessorService
-        ? await detectTraceContext(runtimeContext.executionOptions.traceProcessorService, traceId)
-        : undefined;
-      strategyMatchResult = await this.strategyRegistry.matchEnhanced(
-        query,
-        runtimeContext.intent,
-        traceContext
-      );
-    } catch {
-      strategyMatchResult = null;
-    }
-
-    strategyMatchResult = applyBlockedStrategyIds(
-      strategyMatchResult,
-      runtimeContext.executionOptions.blockedStrategyIds
-    );
-
-    const preferredLoopMode = runtimeContext.sessionContext.getTraceAgentState()?.preferences?.defaultLoopMode;
-    const preferHypothesisLoop = strategyMatchResult?.strategy
-      ? shouldPreferHypothesisLoop({
-          strategyId: strategyMatchResult.strategy.id,
-          preferredLoopMode,
-        })
-      : false;
 
     const incrementalScope = this.determineIncrementalScope(query, runtimeContext.sessionContext);
     emitter.emitUpdate('incremental_scope', {
@@ -352,62 +251,48 @@ export class AgentRuntime extends EventEmitter {
       relevantAgents: incrementalScope.relevantAgents,
     });
 
+    const selection = await selectInitialExecutor({
+      query,
+      traceId,
+      runtimeContext,
+      services,
+      emitter,
+      runtimeConfig: this.runtimeConfig,
+      modelRouter: this.modelRouter,
+      agentRegistry: this.agentRegistry,
+      strategyPlanner: this.strategyPlanner,
+      strategyRegistry: this.strategyRegistry,
+      focusStore: this.focusStore,
+    });
+
     const executionCtx: ExecutionContext = {
       query,
       sessionId,
       traceId,
       intent: runtimeContext.intent,
-      initialHypotheses,
+      initialHypotheses: selection.initialHypotheses,
       sharedContext,
       options: runtimeContext.executionOptions,
       sessionContext: runtimeContext.sessionContext,
       incrementalScope,
-      config: effectiveConfig,
+      config: selection.effectiveConfig,
     };
 
-    if (strategyMatchResult?.strategy) {
+    if (selection.strategyMatchResult?.strategy) {
       executionCtx.options.suggestedStrategy = {
-        id: strategyMatchResult.strategy.id,
-        name: strategyMatchResult.strategy.name,
-        confidence: strategyMatchResult.confidence,
-        matchMethod: strategyMatchResult.matchMethod,
-        reasoning: strategyMatchResult.reasoning,
+        id: selection.strategyMatchResult.strategy.id,
+        name: selection.strategyMatchResult.strategy.name,
+        confidence: selection.strategyMatchResult.confidence,
+        matchMethod: selection.strategyMatchResult.matchMethod,
+        reasoning: selection.strategyMatchResult.reasoning,
       };
     }
 
-    let executor: AnalysisExecutor;
-    if (strategyMatchResult?.strategy && !preferHypothesisLoop) {
-      emitter.emitUpdate('strategy_selected', {
-        strategyId: strategyMatchResult.strategy.id,
-        strategyName: strategyMatchResult.strategy.name,
-        confidence: strategyMatchResult.confidence,
-        reasoning: strategyMatchResult.reasoning || 'keyword match',
-        selectionMethod: strategyMatchResult.matchMethod === 'keyword' ? 'keyword' : 'llm',
-      });
-      executor = new StrategyExecutor(strategyMatchResult.strategy, services);
-    } else {
-      if (strategyMatchResult?.fallbackReason) {
-        emitter.emitUpdate('strategy_fallback', {
-          reason: strategyMatchResult.fallbackReason,
-          candidatesEvaluated: this.strategyRegistry.getAll().length,
-          topCandidateConfidence: strategyMatchResult.confidence,
-          fallbackTo: 'hypothesis_driven',
-        });
-      }
-      const hypothesisExecutor = new HypothesisExecutor(
-        services,
-        this.agentRegistry,
-        this.strategyPlanner
-      );
-      hypothesisExecutor.setFocusStore(this.focusStore);
-      executor = hypothesisExecutor;
-    }
-
     const startTime = Date.now();
-    const executorResult = await executor.execute(executionCtx, emitter);
+    const executorResult = await selection.executor.execute(executionCtx, emitter);
 
-    this.handleExecutorIntervention(sessionId, executorResult);
-    this.applyEntityWriteback(runtimeContext.sessionContext, executorResult);
+    this.resultFinalizer.handleExecutorIntervention(sessionId, executorResult);
+    this.resultFinalizer.applyEntityWriteback(runtimeContext.sessionContext, executorResult);
 
     const previousFindings = runtimeContext.sessionContext.getAllFindings();
     const mergedFindings = incrementalScope.isExtension
@@ -419,68 +304,26 @@ export class AgentRuntime extends EventEmitter {
       : mergedFindings.length > 12
         ? 500
         : 600;
-    const conclusion = await generateConclusion(
-      sharedContext,
-      mergedFindings,
-      runtimeContext.intent,
-      this.modelRouter,
-      emitter,
-      executorResult.stopReason || undefined,
-      {
-        turnCount: runtimeContext.sessionContext.getAllTurns().length,
-        historyContext: runtimeContext.sessionContext.generatePromptContext(conclusionHistoryBudget),
-      }
-    );
 
-    const result: AgentRuntimeAnalysisResult = {
+    return this.resultFinalizer.finalizeAnalysisResult({
+      query,
       sessionId,
-      success: true,
-      findings: mergedFindings,
-      hypotheses: Array.from(sharedContext.hypotheses.values()),
-      conclusion,
-      conclusionContract: deriveConclusionContract(conclusion, {
-        mode: runtimeContext.sessionContext.getAllTurns().length > 0 ? 'focused_answer' : 'initial_report',
-        singleFrameDrillDown: false,
-        sceneId: this.resolveSceneIdHint(runtimeContext.intent, mergedFindings),
-      }) || undefined,
-      confidence: executorResult.confidence,
-      rounds: executorResult.rounds,
-      totalDurationMs: Date.now() - startTime,
-    };
-
-    const recordedTurn = runtimeContext.sessionContext.addTurn(
-      query,
-      runtimeContext.intent,
-      {
-        success: true,
-        findings: executorResult.findings,
-        confidence: result.confidence,
-        message: conclusion,
-      },
-      executorResult.findings
-    );
-    runtimeContext.sessionContext.updateWorkingMemoryFromConclusion({
-      turnIndex: recordedTurn.turnIndex,
-      query,
-      conclusion,
-      confidence: result.confidence,
+      intent: runtimeContext.intent,
+      sessionContext: runtimeContext.sessionContext,
+      sharedContext,
+      emitter,
+      executorResult,
+      mergedFindings,
+      startTime,
+      singleFrameDrillDown: false,
+      mode: runtimeContext.sessionContext.getAllTurns().length > 0 ? 'focused_answer' : 'initial_report',
+      historyBudget: conclusionHistoryBudget,
     });
-    runtimeContext.sessionContext.recordTraceAgentTurn({
-      turnId: recordedTurn.id,
-      turnIndex: recordedTurn.turnIndex,
-      query,
-      followUpType: runtimeContext.intent.followUpType,
-      intentPrimaryGoal: runtimeContext.intent.primaryGoal,
-      conclusion,
-      confidence: result.confidence,
-    });
-
-    return result;
   }
 
   private determineIncrementalScope(
     query: string,
-    sessionContext: EnhancedSessionContext
+    sessionContext: ReturnType<typeof sessionContextManager.getOrCreate>
   ): IncrementalScope {
     const entityStore = sessionContext.getEntityStore();
     const previousFindings = sessionContext.getAllFindings();
@@ -510,7 +353,7 @@ export class AgentRuntime extends EventEmitter {
     traceId: string,
     runtimeContext: PreparedRuntimeContext
   ): Promise<AgentRuntimeAnalysisResult> {
-    const services = this.createNativeExecutionServices();
+    const services = this.executionFactory.createExecutionServices();
     const emitter = this.createRuntimeEmitter();
     const sharedContext = services.messageBus.createSharedContext(sessionId, traceId);
 
@@ -526,7 +369,7 @@ export class AgentRuntime extends EventEmitter {
       config: this.runtimeConfig,
     };
 
-    const executor = this.createModeExecutor(runtimeContext, services);
+    const executor = this.executionFactory.createFollowUpModeExecutor(runtimeContext, services);
     if (!executor) {
       return {
         sessionId,
@@ -543,194 +386,34 @@ export class AgentRuntime extends EventEmitter {
     const startTime = Date.now();
     const executorResult = await executor.execute(executionCtx, emitter);
 
-    this.handleExecutorIntervention(sessionId, executorResult);
-    this.applyEntityWriteback(runtimeContext.sessionContext, executorResult);
+    this.resultFinalizer.handleExecutorIntervention(sessionId, executorResult);
+    this.resultFinalizer.applyEntityWriteback(runtimeContext.sessionContext, executorResult);
 
     const previousFindings = runtimeContext.sessionContext.getAllFindings();
     const mergedFindings = runtimeContext.decisionContext.mode === 'extend'
       ? this.incrementalAnalyzer.mergeFindings(previousFindings, executorResult.findings)
       : executorResult.findings;
 
-    emitter.emitUpdate('progress', { phase: 'concluding', message: '生成分析结论' });
-    const conclusion = await generateConclusion(
-      sharedContext,
-      mergedFindings,
-      runtimeContext.intent,
-      this.modelRouter,
-      emitter,
-      executorResult.stopReason || undefined,
-      {
-        turnCount: runtimeContext.sessionContext.getAllTurns().length,
-        historyContext: runtimeContext.sessionContext.generatePromptContext(600),
-      }
-    );
-
     const singleFrameDrillDown =
       runtimeContext.intent.followUpType === 'drill_down' &&
       (runtimeContext.intent.referencedEntities || []).filter(entity => entity.type === 'frame').length === 1;
 
-    const result: AgentRuntimeAnalysisResult = {
+    emitter.emitUpdate('progress', { phase: 'concluding', message: '生成分析结论' });
+
+    return this.resultFinalizer.finalizeAnalysisResult({
+      query,
       sessionId,
-      success: true,
-      findings: mergedFindings,
-      hypotheses: Array.from(sharedContext.hypotheses.values()),
-      conclusion,
-      conclusionContract: deriveConclusionContract(conclusion, {
-        mode: runtimeContext.sessionContext.getAllTurns().length > 0 ? 'focused_answer' : 'initial_report',
-        singleFrameDrillDown,
-        sceneId: this.resolveSceneIdHint(runtimeContext.intent, mergedFindings),
-      }) || undefined,
-      confidence: executorResult.confidence,
-      rounds: executorResult.rounds,
-      totalDurationMs: Date.now() - startTime,
-    };
-
-    const recordedTurn = runtimeContext.sessionContext.addTurn(
-      query,
-      runtimeContext.intent,
-      {
-        success: true,
-        findings: executorResult.findings,
-        confidence: result.confidence,
-        message: conclusion,
-      },
-      executorResult.findings
-    );
-    runtimeContext.sessionContext.updateWorkingMemoryFromConclusion({
-      turnIndex: recordedTurn.turnIndex,
-      query,
-      conclusion,
-      confidence: result.confidence,
+      intent: runtimeContext.intent,
+      sessionContext: runtimeContext.sessionContext,
+      sharedContext,
+      emitter,
+      executorResult,
+      mergedFindings,
+      startTime,
+      singleFrameDrillDown,
+      mode: runtimeContext.sessionContext.getAllTurns().length > 0 ? 'focused_answer' : 'initial_report',
+      historyBudget: 600,
     });
-    runtimeContext.sessionContext.recordTraceAgentTurn({
-      turnId: recordedTurn.id,
-      turnIndex: recordedTurn.turnIndex,
-      query,
-      followUpType: runtimeContext.intent.followUpType,
-      intentPrimaryGoal: runtimeContext.intent.primaryGoal,
-      conclusion,
-      confidence: result.confidence,
-    });
-
-    return result;
-  }
-
-  private createNativeExecutionServices(): AnalysisServices {
-    const messageBus = createAgentMessageBus({
-      maxConcurrentTasks: this.runtimeConfig.maxConcurrentTasks,
-      messageTimeoutMs: this.runtimeConfig.taskTimeoutMs ?? DEFAULT_CONFIG.taskTimeoutMs ?? 180000,
-      enableLogging: this.runtimeConfig.enableLogging,
-    });
-
-    for (const agent of this.agentRegistry.getAll()) {
-      messageBus.registerAgent(agent);
-    }
-
-    return {
-      modelRouter: this.modelRouter,
-      messageBus,
-      circuitBreaker: this.circuitBreaker,
-      emittedEnvelopeRegistry: createEmittedEnvelopeRegistry(),
-    };
-  }
-
-  private createModeExecutor(
-    runtimeContext: PreparedRuntimeContext,
-    services: AnalysisServices
-  ): AnalysisExecutor | null {
-    if (runtimeContext.decisionContext.mode === 'compare') {
-      return new ComparisonExecutor(
-        runtimeContext.sessionContext,
-        services,
-        runtimeContext.executionOptions.traceProcessorService,
-        runtimeContext.sessionContext.getTraceId()
-      );
-    }
-
-    if (runtimeContext.decisionContext.mode === 'extend') {
-      const extendExecutor = new ExtendExecutor(
-        runtimeContext.sessionContext,
-        services,
-        runtimeContext.executionOptions.traceProcessorService,
-        runtimeContext.sessionContext.getTraceId()
-      );
-      extendExecutor.setFocusStore(this.focusStore);
-      return extendExecutor;
-    }
-
-    if (runtimeContext.decisionContext.mode === 'drill_down') {
-      const intervals = runtimeContext.executionOptions.prebuiltIntervals || runtimeContext.followUp.focusIntervals || [];
-      if (intervals.length === 0) {
-        return null;
-      }
-      const followUp: FollowUpResolution = {
-        ...runtimeContext.followUp,
-        focusIntervals: intervals,
-      };
-      return new DirectDrillDownExecutor(followUp, services);
-    }
-
-    return null;
-  }
-
-  private handleExecutorIntervention(sessionId: string, executorResult: ExecutorResult): void {
-    if (!executorResult.interventionRequest) {
-      return;
-    }
-
-    const intervention = executorResult.interventionRequest;
-    const options = [
-      {
-        id: 'continue',
-        label: '继续分析',
-        description: '继续当前分析策略',
-        action: 'continue' as const,
-        recommended: true,
-      },
-      {
-        id: 'abort',
-        label: '结束分析',
-        description: '以当前结果结束',
-        action: 'abort' as const,
-      },
-    ];
-
-    this.interventionController.createAgentIntervention(
-      sessionId,
-      intervention.reason,
-      options,
-      {
-        currentFindings: executorResult.findings,
-        possibleDirections: intervention.possibleDirections.map(direction => ({
-          id: direction.id,
-          description: direction.description,
-          confidence: direction.confidence,
-          requiredAgents: [],
-        })),
-        elapsedTimeMs: intervention.elapsedTimeMs,
-        confidence: intervention.confidence,
-        roundsCompleted: intervention.roundsCompleted,
-        progressSummary: intervention.progressSummary,
-      }
-    );
-  }
-
-  private applyEntityWriteback(sessionContext: EnhancedSessionContext, executorResult: ExecutorResult): void {
-    if (executorResult.capturedEntities) {
-      applyCapturedEntities(sessionContext.getEntityStore(), executorResult.capturedEntities);
-    }
-
-    if (executorResult.analyzedEntityIds) {
-      const store = sessionContext.getEntityStore();
-      for (const frameId of executorResult.analyzedEntityIds.frames || []) {
-        store.markFrameAnalyzed(frameId);
-      }
-      for (const sessionId of executorResult.analyzedEntityIds.sessions || []) {
-        store.markSessionAnalyzed(sessionId);
-      }
-    }
-
-    sessionContext.refreshTraceAgentCoverage();
   }
 
   private async executeNativeClarify(
@@ -888,14 +571,6 @@ export class AgentRuntime extends EventEmitter {
   }
 }
 
-interface PreparedRuntimeContext {
-  sessionContext: EnhancedSessionContext;
-  intent: Intent;
-  followUp: FollowUpResolution;
-  decisionContext: DecisionContext;
-  executionOptions: AnalysisOptions;
-}
-
 export function createAgentRuntime(
   modelRouter: ModelRouter,
   config?: Partial<AgentRuntimeConfig>
@@ -903,198 +578,15 @@ export function createAgentRuntime(
   return new AgentRuntime(modelRouter, config);
 }
 
-export function buildDecisionContextFromIntent(
-  query: string,
-  sessionContext: EnhancedSessionContext,
-  intent: Intent,
-  followUpResolution: { isFollowUp: boolean; confidence: number; resolvedParams: Record<string, unknown> },
-  focusIntervals: Array<{ startTs: string; endTs: string }>
-): DecisionContext {
-  const traceAgentState = sessionContext.getTraceAgentState();
-  const turns = sessionContext.getAllTurns();
-  const mode = mapFollowUpTypeToMode(intent.followUpType);
-  const requestedDomains = deriveRequestedDomainsFromIntent(intent, query);
-  const requestedActions = deriveRequestedActionsFromIntent(intent, followUpResolution, focusIntervals);
-
-  return {
-    sessionId: sessionContext.getSessionId(),
-    traceId: sessionContext.getTraceId(),
-    turnIndex: turns.length,
-    mode,
-    userGoal: intent.primaryGoal || query,
-    requestedDomains,
-    requestedActions,
-    referencedEntities: mapReferencedEntities(intent.referencedEntities || []),
-    coverageDomains: traceAgentState?.coverage?.domains || [],
-    evidenceCount: Array.isArray(traceAgentState?.evidence) ? traceAgentState!.evidence.length : 0,
-    contradictionCount: Array.isArray(traceAgentState?.contradictions)
-      ? traceAgentState!.contradictions.length
-      : 0,
-  };
-}
-
-export function buildRuntimeExecutionOptions(
-  baseOptions: AnalysisOptions,
-  followUpResolution: {
-    resolvedParams: Record<string, unknown>;
-    confidence: number;
-    focusIntervals?: FocusInterval[];
-  },
-  resolvedIntervals: FocusInterval[] | undefined,
-  intent: Intent
-): AnalysisOptions {
-  return {
-    ...baseOptions,
-    ...(Object.keys(followUpResolution.resolvedParams || {}).length > 0
-      ? { resolvedFollowUpParams: followUpResolution.resolvedParams }
-      : {}),
-    ...(Array.isArray(resolvedIntervals) && resolvedIntervals.length > 0
-      ? { prebuiltIntervals: resolvedIntervals }
-      : {}),
-    ...(intent.followUpType === 'drill_down' ? { suggestedStrategy: { id: 'drill_down', name: 'Direct drill-down', confidence: followUpResolution.confidence } } : {}),
-  };
-}
-
-export function applyBlockedStrategyIds(
-  matchResult: StrategyMatchResult | null,
-  blockedStrategyIds?: string[]
-): StrategyMatchResult | null {
-  if (!matchResult?.strategy) {
-    return matchResult;
-  }
-
-  const blocked = new Set(
-    (blockedStrategyIds || [])
-      .map(id => String(id || '').trim())
-      .filter(Boolean)
-  );
-  if (!blocked.has(matchResult.strategy.id)) {
-    return matchResult;
-  }
-
-  return {
-    strategy: null,
-    matchMethod: matchResult.matchMethod,
-    confidence: matchResult.confidence,
-    reasoning: matchResult.reasoning,
-    shouldFallback: true,
-    fallbackReason: `策略 ${matchResult.strategy.id} 已被 blockedStrategyIds 禁用`,
-  };
-}
-
-export function mapFollowUpTypeToMode(followUpType: Intent['followUpType']): DecisionContext['mode'] {
-  if (followUpType === 'clarify') return 'clarify';
-  if (followUpType === 'compare') return 'compare';
-  if (followUpType === 'extend') return 'extend';
-  if (followUpType === 'drill_down') return 'drill_down';
-  return 'initial';
-}
-
-export function deriveRequestedDomainsFromIntent(intent: Intent, query: string): string[] {
-  const aspectTokens = Array.isArray(intent.aspects)
-    ? intent.aspects.map(token => String(token || '').toLowerCase())
-    : [];
-  const queryTokens = String(query || '').toLowerCase();
-
-  const mappings: Array<{ tokens: string[]; domain: string }> = [
-    { tokens: ['frame', 'jank', 'render', '卡顿', '帧'], domain: 'frame' },
-    { tokens: ['cpu', 'sched', '调度'], domain: 'cpu' },
-    { tokens: ['binder', 'ipc'], domain: 'binder' },
-    { tokens: ['memory', 'gc', '内存'], domain: 'memory' },
-    { tokens: ['startup', 'launch', '启动'], domain: 'startup' },
-    { tokens: ['gpu'], domain: 'gpu' },
-    { tokens: ['surfaceflinger', 'sf'], domain: 'surfaceflinger' },
-    { tokens: ['input', 'touch', '交互'], domain: 'interaction' },
-  ];
-
-  const domains = mappings
-    .filter(item =>
-      item.tokens.some(token =>
-        aspectTokens.some(aspect => aspect.includes(token)) || queryTokens.includes(token)
-      )
-    )
-    .map(item => item.domain);
-
-  return domains.length > 0 ? Array.from(new Set(domains)) : ['frame', 'cpu'];
-}
-
-function deriveRequestedActionsFromIntent(
-  intent: Intent,
-  followUpResolution: { isFollowUp: boolean },
-  focusIntervals: Array<{ startTs: string; endTs: string }>
-): string[] {
-  const actions: string[] = [];
-  if (intent.followUpType === 'compare') actions.push('compare_entities');
-  if (intent.followUpType === 'extend') actions.push('expand_scope');
-  if (intent.followUpType === 'drill_down') actions.push('drill_down');
-  if (followUpResolution.isFollowUp) actions.push('follow_up');
-  if (focusIntervals.length > 0) actions.push('has_focus_intervals');
-  return actions;
-}
-
-function mapReferencedEntities(referencedEntities: ReferencedEntity[]): DecisionContext['referencedEntities'] {
-  const allowedTypes = new Set<DecisionContext['referencedEntities'][number]['type']>([
-    'frame',
-    'session',
-    'startup',
-    'process',
-    'binder_call',
-    'time_range',
-  ]);
-
-  const mapped: DecisionContext['referencedEntities'] = [];
-  for (const entity of referencedEntities) {
-    if (!allowedTypes.has(entity.type as DecisionContext['referencedEntities'][number]['type'])) {
-      continue;
-    }
-    mapped.push({
-      type: entity.type as DecisionContext['referencedEntities'][number]['type'],
-      id: entity.id,
-      value: entity.value,
-    });
-  }
-
-  return mapped;
-}
-
-export function buildNativeClarifyPrompt(
-  query: string,
-  contextSummary: string,
-  recentFindings: Finding[]
-): string {
-  const parts: string[] = [];
-  parts.push('你是 SmartPerfetto 的 Android 性能分析助手。请回答用户的澄清问题。');
-  parts.push('要求：只基于给定上下文，不编造数据；若信息不足，明确说明不足。');
-  parts.push('');
-  parts.push(`用户问题: ${query}`);
-  parts.push('');
-
-  if (contextSummary.trim()) {
-    parts.push('上下文摘要:');
-    parts.push(contextSummary);
-    parts.push('');
-  }
-
-  if (recentFindings.length > 0) {
-    parts.push('近期发现:');
-    for (const finding of recentFindings) {
-      parts.push(`- [${finding.severity}] ${finding.title}: ${finding.description}`);
-    }
-    parts.push('');
-  }
-
-  parts.push('请直接给出中文解释，结构：结论 -> 依据 -> 建议（如果有）。');
-  return parts.join('\n');
-}
-
-export function buildNativeClarifyFallback(query: string, recentFindings: Finding[]): string {
-  if (recentFindings.length === 0) {
-    return `当前缺少足够上下文来直接回答“${query}”。建议先运行一次完整分析，再针对具体帧/会话继续提问。`;
-  }
-
-  const top = recentFindings[0];
-  return `基于当前会话，最相关发现是“${top.title}”。\n${top.description}\n如果你希望，我可以继续展开这个发现的根因链路和优化优先级。`;
-}
+export {
+  applyBlockedStrategyIds,
+  buildDecisionContextFromIntent,
+  buildNativeClarifyFallback,
+  buildNativeClarifyPrompt,
+  buildRuntimeExecutionOptions,
+  deriveRequestedDomainsFromIntent,
+  mapFollowUpTypeToMode,
+} from './runtimeContextBuilder';
 
 function buildPrinciplesAppliedUpdate(decision: PrincipleDecision, planId: string): StreamingUpdate {
   return {
